+++ packages/core/demo/demo-data/line.ts
@@ -1,4 +1,3 @@
-import { colors } from "./colors";
 import { Tools } from "../../src/tools";
 
 export const lineTimeSeriesData = {
@@ -6,7 +5,6 @@ export const lineTimeSeriesData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			fillColors: [colors[0]],
 			data: [
 				{
 					date: new Date(2019, 0, 1),
@@ -32,7 +30,6 @@ export const lineTimeSeriesData = {
 		},
 		{
 			label: "Dataset 2",
-			fillColors: [colors[1]],
 			data: [
 				{
 					date: new Date(2019, 0, 2),
@@ -58,7 +55,6 @@ export const lineTimeSeriesData = {
 		},
 		{
 			label: "Dataset 3",
-			fillColors: [colors[2]],
 			data: [
 				{
 					date: new Date(2019, 0, 1),
@@ -84,7 +80,6 @@ export const lineTimeSeriesData = {
 		},
 		{
 			label: "Dataset 4",
-			fillColors: [colors[3]],
 			data: [
 				{
 					date: new Date(2019, 0, 2),
@@ -130,7 +125,6 @@ export const lineData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			fillColors: [colors[0]],
 			data: [
 				32100,
 				23500,
@@ -141,7 +135,6 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 2",
-			fillColors: [colors[1]],
 			data: [
 				34200,
 				53200,
@@ -152,7 +145,6 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 3 long name",
-			fillColors: [colors[2]],
 			data: [
 				41200,
 				23400,
@@ -163,7 +155,6 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 4 long name",
-			fillColors: [colors[3]],
 			data: [
 				22000,
 				1200,
@@ -174,7 +165,6 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 5 long name",
-			fillColors: [colors[4]],
 			data: [
 				2412,
 				30000,
@@ -185,7 +175,6 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 6 long name",
-			fillColors: [colors[5]],
 			data: [
 				0,
 				20000,
@@ -232,7 +221,6 @@ export const scatterData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			fillColors: [colors[0]],
 			data: [
 				32100,
 				23500,
@@ -243,7 +231,6 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 2",
-			fillColors: [colors[1]],
 			data: [
 				34200,
 				53200,
@@ -254,7 +241,6 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 3 long name",
-			fillColors: [colors[2]],
 			data: [
 				41200,
 				23400,
@@ -265,7 +251,6 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 4 long name",
-			fillColors: [colors[3]],
 			data: [
 				22000,
 				1200,
@@ -276,7 +261,6 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 5 long name",
-			fillColors: [colors[4]],
 			data: [
 				2412,
 				30000,
@@ -287,7 +271,6 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 6 long name",
-			fillColors: [colors[5]],
 			data: [
 				0,
 				20000,