+++ packages/core/src/charts/bar-grouped.ts
@@ -12,7 +12,7 @@ import {
 	Grid,
 	GroupedBar,
 	TwoDimensionalAxes,
-	HorizontalZeroLine,
+	ZeroLine,
 	TooltipBar,
 	// the imports below are needed because of typescript bug (error TS4029)
 	Tooltip,
@@ -43,7 +43,7 @@ export class GroupedBarChart extends AxisChart {
 			new TwoDimensionalAxes(this.model, this.services),
 			new Grid(this.model, this.services),
 			new GroupedBar(this.model, this.services),
-			new HorizontalZeroLine(this.model, this.services)
+			new ZeroLine(this.model, this.services)
 		];
 
 		const components: any[] = this.getAxisChartComponents(graphFrameComponents);