+++ packages/core/src/components/graphs/bar-stacked.ts
@@ -1,7 +1,13 @@
 // Internal Imports
-import { ScaleTypes, TooltipTypes } from "../../interfaces/enums";
 import { Tools } from "../../tools";
 import { Bar } from "./bar";
+import {
+	Roles,
+	ScaleTypes,
+	TooltipTypes,
+	Events,
+	CartesianOrientations
+} from "../../interfaces";
 
 // D3 Imports
 import { select } from "d3-selection";
@@ -39,7 +45,11 @@ export class StackedBar extends Bar {
 		let stackDataArray;
 		const displayData = this.model.getDisplayData();
 
-		const timeSeries = this.services.axes.getMainXAxis().scaleType === ScaleTypes.TIME;
+		// the domain axis for stack data depends on the orientation of the bar chart
+		const domainAxisPosition = this.services.cartesianScales.getDomainAxisPosition();
+		const domainScaleType = this.services.cartesianScales.getScaleTypeByPosition(domainAxisPosition);
+		// get scale type for the main axis
+		const timeSeries = domainScaleType === ScaleTypes.TIME;
 
 		if (timeSeries) {
 			// Get all date values provided in data
@@ -111,42 +121,73 @@ export class StackedBar extends Bar {
 		// Add bar groups that need to be introduced
 		barGroups.enter()
 			.append("g")
-			.classed("bars", true);
+			.classed("bars", true)
+			.attr("role", Roles.GROUP);
 
 		// Update data on all bars
-		const bars = svg.selectAll("g.bars").selectAll("rect.bar")
+		const bars = svg.selectAll("g.bars").selectAll("path.bar")
 			.data(d => addLabelsAndValueToData(d), d => d.label);
 
 		// Remove bars that need to be removed
 		bars.exit()
 			.remove();
 
-		// Update styling and position on existing bars
-		// As well as bars that were just added
+		const yScale = this.services.cartesianScales.getRangeScale();
 		bars.enter()
-			.append("rect")
+			.append("path")
 			.merge(bars)
 				.classed("bar", true)
-				.attr("x", (d, i) => {
-					const barWidth = this.getBarWidth();
-					return this.services.axes.getXValue(d, i) - (barWidth / 2);
-				})
-				.attr("width", this.getBarWidth.bind(this))
 				.transition(this.services.transitions.getTransition("bar-update-enter", animate))
-				.attr("y", (d, i) => this.services.axes.getYValue(d[1], i))
 				.attr("fill", d => this.model.getFillScale()[d.datasetLabel](d.label))
 				.attr("height", (d, i) => {
 					const { datasetLabel } = d;
 					const datasetLabelIndex = stackKeys.indexOf(datasetLabel);
-					const height = this.services.axes.getYValue(d[0]) - this.services.axes.getYValue(d[1]);
-
+					let height;
+					// determine height based on the y axis
+					if (yScale.scaleType === ScaleTypes.LABELS) {
+						height = Math.abs(yScale.scale.range()[0] - this.services.cartesianScales.getRangeValue(d.label, i));
+					} else {
+						height = this.services.cartesianScales.getRangeValue(d[0]) - this.services.cartesianScales.getRangeValue(d[1]);
+					}
+					// create dividers between bars
 					if (datasetLabelIndex > 0 && height >= options.bars.dividerSize) {
 						return height - options.bars.dividerSize;
 					}
 
 					return height;
 				})
-				.attr("opacity", 1);
+				.attr("d", (d, i) => {
+					/*
+					* Orientation support for horizontal/vertical bar charts
+					* Determine coordinates needed for a vertical set of paths
+					* to draw the bars needed, and pass those coordinates down to
+					* generateSVGPathString() to decide whether it needs to flip them
+					*/
+					const barWidth = this.getBarWidth();
+					const x0 = this.services.cartesianScales.getDomainValue(d, i) - barWidth / 2;
+					const x1 = x0 + barWidth;
+					const y0 = this.services.cartesianScales.getRangeValue(d[0], i);
+					let y1 = this.services.cartesianScales.getRangeValue(d[1], i);
+
+					// Add the divider gap
+					if (Math.abs(y1 - y0) > 0 && Math.abs(y1 - y0) > options.bars.dividerSize) {
+						if (this.services.cartesianScales.getOrientation() === CartesianOrientations.VERTICAL) {
+							y1 += 1;
+						} else {
+							y1 -= 1;
+						}
+					}
+
+					return Tools.generateSVGPathString(
+						{ x0, x1, y0, y1 },
+						this.services.cartesianScales.getOrientation()
+					);
+				})
+				.attr("opacity", 1)
+				// a11y
+				.attr("role", Roles.GRAPHICS_SYMBOL)
+				.attr("aria-roledescription", "bar")
+				.attr("aria-label", d => d.value);
 
 		// Add event listeners for the above elements
 		this.addEventListeners();
@@ -156,28 +197,34 @@ export class StackedBar extends Bar {
 	handleLegendOnHover = (event: CustomEvent) => {
 		const { hoveredElement } = event.detail;
 
-		this.parent.selectAll("rect.bar")
+		this.parent.selectAll("path.bar")
 			.transition(this.services.transitions.getTransition("legend-hover-bar"))
 			.attr("opacity", d => (d.datasetLabel !== hoveredElement.datum()["key"]) ? 0.3 : 1);
 	}
 
 	// Un-highlight all elements
 	handleLegendMouseOut = (event: CustomEvent)  => {
-		this.parent.selectAll("rect.bar")
+		this.parent.selectAll("path.bar")
 			.transition(this.services.transitions.getTransition("legend-mouseout-bar"))
 			.attr("opacity", 1);
 	}
 
 	addEventListeners() {
 		const self = this;
-		this.parent.selectAll("rect.bar")
-			.on("mouseover", function() {
+		this.parent.selectAll("path.bar")
+			.on("mouseover", function(datum) {
 				const hoveredElement = select(this);
 
 				hoveredElement.transition(self.services.transitions.getTransition("graph_element_mouseover_fill_update"))
 					.attr("fill", color(hoveredElement.attr("fill")).darker(0.7).toString());
+
+				// Dispatch mouse event
+				self.services.events.dispatchEvent(Events.Bar.BAR_MOUSEOVER, {
+					element: hoveredElement,
+					datum
+				});
 			})
-			.on("mousemove", function() {
+			.on("mousemove", function(datum) {
 				const hoveredElement = select(this);
 				const itemData = select(this).datum();
 				hoveredElement.classed("hovered", true);
@@ -196,6 +243,12 @@ export class StackedBar extends Bar {
 						label: sharedLabel
 					}));
 
+				// Dispatch mouse event
+				self.services.events.dispatchEvent(Events.Bar.BAR_MOUSEMOVE, {
+					element: hoveredElement,
+					datum
+				});
+
 				// Show tooltip
 				self.services.events.dispatchEvent("show-tooltip", {
 					multidata: activePoints,
@@ -203,12 +256,25 @@ export class StackedBar extends Bar {
 					type: TooltipTypes.DATAPOINT
 				});
 			})
-			.on("mouseout", function() {
+			.on("click", function(datum) {
+				// Dispatch mouse event
+				self.services.events.dispatchEvent(Events.Bar.BAR_CLICK, {
+					element: select(this),
+					datum
+				});
+			})
+			.on("mouseout", function(datum) {
 				const hoveredElement = select(this);
 				hoveredElement.classed("hovered", false);
 
 				hoveredElement.transition(self.services.transitions.getTransition("graph_element_mouseout_fill_update"))
 					.attr("fill", (d: any) => self.model.getFillScale()[d.datasetLabel](d.label));
+
+				// Dispatch mouse event
+				self.services.events.dispatchEvent(Events.Bar.BAR_MOUSEOUT, {
+					element: hoveredElement,
+					datum
+				});
 
 				// Hide tooltip
 				self.services.events.dispatchEvent("hide-tooltip", { hoveredElement });
@@ -217,7 +283,7 @@ export class StackedBar extends Bar {
 
 	destroy() {
 		// Remove event listeners
-		this.parent.selectAll("rect.bar")
+		this.parent.selectAll("path.bar")
 			.on("mouseover", null)
 			.on("mousemove", null)
 			.on("mouseout", null);