+++ packages/core/src/components/graphs/scatter.ts
@@ -1,9 +1,9 @@
 // Internal Imports
 import { Component } from "../component";
+import { TooltipTypes, Roles, Events } from "../../interfaces";
 
 // D3 Imports
-import { select } from "d3-selection";
-import { TooltipTypes } from "../../interfaces";
+import { select, event as d3Event } from "d3-selection";
 
 export class Scatter extends Component {
 	type = "scatter";
@@ -37,7 +37,8 @@ export class Scatter extends Component {
 		// Add the dot groups that need to be introduced
 		const dotGroupsEnter = dotGroups.enter()
 			.append("g")
-				.classed("dots", true);
+				.classed("dots", true)
+				.attr("role", Roles.GROUP);
 
 		// Update data on all circles
 		const dots = dotGroupsEnter.merge(dotGroups)
@@ -56,9 +57,9 @@ export class Scatter extends Component {
 			.classed("dot", true)
 			.classed("filled", filled)
 			.classed("unfilled", !filled)
-			.attr("cx", (d, i) => this.services.axes.getXValue(d, i))
+			.attr("cx", (d, i) => this.services.cartesianScales.getDomainValue(d, i))
 			.transition(this.services.transitions.getTransition("scatter-update-enter", animate))
-			.attr("cy", (d, i) => this.services.axes.getYValue(d, i))
+			.attr("cy", (d, i) => this.services.cartesianScales.getRangeValue(d, i))
 			.attr("r", options.points.radius)
 			.attr("fill", d => {
 				if (filled) {
@@ -67,7 +68,11 @@ export class Scatter extends Component {
 			})
 			.attr("fill-opacity", filled ? 0.2 : 1)
 			.attr("stroke", d => this.model.getStrokeColor(d.datasetLabel, d.label, d.value))
-			.attr("opacity", 1);
+			.attr("opacity", 1)
+			// a11y
+			.attr("role", Roles.GRAPHICS_SYMBOL)
+			.attr("aria-roledescription", "point")
+			.attr("aria-label", d => d.value);
 
 		// Add event listeners to elements drawn
 		this.addEventListeners();
@@ -102,11 +107,18 @@ export class Scatter extends Component {
 	addEventListeners() {
 		const self = this;
 		this.parent.selectAll("circle")
-			.on("mouseover mousemove", function() {
+			.on("mouseover mousemove", function(datum) {
 				const hoveredElement = select(this);
 				hoveredElement.classed("hovered", true);
 
 				hoveredElement.style("fill", (d: any) => self.model.getFillScale()[d.datasetLabel](d.label));
+
+				const eventNameToDispatch = d3Event.type === "mouseover" ? Events.Scatter.SCATTER_MOUSEOVER : Events.Scatter.SCATTER_MOUSEMOVE;
+				// Dispatch mouse event
+				self.services.events.dispatchEvent(eventNameToDispatch, {
+					element: hoveredElement,
+					datum
+				});
 
 				// Show tooltip
 				self.services.events.dispatchEvent("show-tooltip", {
@@ -114,13 +126,26 @@ export class Scatter extends Component {
 					type: TooltipTypes.DATAPOINT
 				});
 			})
-			.on("mouseout", function() {
+			.on("click", function(datum) {
+				// Dispatch mouse event
+				self.services.events.dispatchEvent(Events.Scatter.SCATTER_CLICK, {
+					element: select(this),
+					datum
+				});
+			})
+			.on("mouseout", function(datum) {
 				const hoveredElement = select(this);
 				hoveredElement.classed("hovered", false);
 
 				if (!self.configs.filled) {
 					hoveredElement.style("fill", null);
 				}
+
+				// Dispatch mouse event
+				self.services.events.dispatchEvent(Events.Scatter.SCATTER_MOUSEOUT, {
+					element: hoveredElement,
+					datum
+				});
 
 				// Hide tooltip
 				self.services.events.dispatchEvent("hide-tooltip", { hoveredElement });