+++ packages/core/src/configuration.ts
@@ -18,7 +18,7 @@ import {
 	LegendOptions,
 	ChartTheme,
 	LegendPositions,
-	StackedBarOptions,
+	StackedBarOptions
 } from "./interfaces/index";
 
 /*
@@ -33,6 +33,7 @@ import {
 export const legend: LegendOptions = {
 	position: LegendPositions.BOTTOM,
 	clickable: true,
+	enabled: true,
 	items: {
 		status: {
 			ACTIVE: 1,
@@ -92,10 +93,10 @@ export const barChartTooltip: BarTooltipOptions = Tools.merge({}, axisChartToolt
 
 // We setup no axes by default, the TwoDimensionalAxes component
 // Will setup axes options based on what user provides
-export const axes: AxesOptions = { };
+const axes: AxesOptions = { };
 
-export const timeScale: TimeScaleOptions = {
-	addSpaceOnEdges: true
+const timeScale: TimeScaleOptions = {
+	addSpaceOnEdges: 1,
 };
 
 /**
@@ -105,7 +106,6 @@ const chart: BaseChartOptions = {
 	width: "100%",
 	height: "100%",
 	resizable: true,
-	theme: ChartTheme.DEFAULT,
 	tooltip: baseTooltip,
 	legend,
 	style: {
@@ -131,9 +131,9 @@ const baseBarChart: BarChartOptions = Tools.merge({}, axisChart, {
 		maxWidth: 16
 	},
 	timeScale: Tools.merge(timeScale, {
-		addSpaceOnEdges: true
+		addSpaceOnEdges: 1
 	} as TimeScaleOptions),
-	tooltip: barChartTooltip
+	tooltip: barChartTooltip,
 } as BarChartOptions);
 
 /**
@@ -202,6 +202,9 @@ const pieChart: PieChartOptions = Tools.merge({}, chart, {
 			offsetY: 12,
 			horizontalLineLength: 8,
 			textMargin: 2
+		},
+		labels: {
+			formatter: null
 		}
 	}
 } as PieChartOptions);
@@ -214,7 +217,8 @@ const donutChart: DonutChartOptions = Tools.merge({}, pieChart, {
 		center: {
 			numberFontSize: radius => Math.min((radius / 100) * 24, 24) + "px",
 			titleFontSize: radius => Math.min((radius / 100) * 15, 15) + "px",
-			titleYPosition: radius => Math.min((radius / 80) * 20, 20)
+			titleYPosition: radius => Math.min((radius / 80) * 20, 20),
+			numberFormatter: number => Math.floor(number).toLocaleString()
 		}
 	}
 } as DonutChartOptions);
@@ -259,5 +263,18 @@ export const transitions = {
 	},
 	graph_element_mouseout_fill_update: {
 		duration: 100
+	}
+};
+
+export const axis = {
+	ticks: {
+		number: 7,
+		rotateIfSmallerThan: 30
+	}
+};
+
+export const spacers = {
+	default: {
+		size: 24
 	}
 };