+++ packages/core/src/tools.ts
@@ -1,3 +1,6 @@
+// Internal imports
+import { CartesianOrientations } from "./interfaces";
+
 import {
 	debounce as lodashDebounce,
 	merge as lodashMerge,
@@ -77,23 +80,6 @@ export namespace Tools {
 			x: parseFloat(xyString[0]),
 			y: parseFloat(xyString[1])
 		};
-	}
-
-	/**
-	 * Returns string value for height/width using pixels if there isn't a specified unit of measure
-	 *
-	 * @param value string or number value to be checked for unit of measure
-	 */
-	export function formatWidthHeightValues(value) {
-		const stringValue = value.toString();
-
-		// If the value provided contains any letters
-		// Return it the same way
-		if (stringValue.match(/[a-z]/i)) {
-			return stringValue;
-		}
-
-		return stringValue + "px";
 	}
 
 	/**
@@ -218,5 +204,31 @@ export namespace Tools {
 		}
 
 		return null;
+	};
+
+	interface SVGPathCoordinates {
+		x0: number;
+		x1: number;
+		y0: number;
+		y1: number;
+	}
+
+	export const flipSVGCoordinatesBasedOnOrientation = (verticalCoordinates: SVGPathCoordinates, orientation?: CartesianOrientations) => {
+		if (orientation === CartesianOrientations.HORIZONTAL) {
+			return {
+				y0: verticalCoordinates.x0,
+				y1: verticalCoordinates.x1,
+				x0: verticalCoordinates.y0,
+				x1: verticalCoordinates.y1
+			};
+		}
+
+		return verticalCoordinates;
+	};
+
+	export const generateSVGPathString = (verticalCoordinates: SVGPathCoordinates, orientation?: CartesianOrientations) => {
+		const { x0, x1, y0, y1 } = flipSVGCoordinatesBasedOnOrientation(verticalCoordinates, orientation);
+
+		return `M${x0},${y0}L${x0},${y1}L${x1},${y1}L${x1},${y0}L${x0},${y0}`;
 	};
 }